alterado estrutura de pastas para possibilitar melhor reuso de código
