--- src/Authentication/JWT/Exceptions/NotPermissionAccessException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Bludata\Authentication\JWT\Exceptions;
-
-final class NotPermissionAccessException extends \Exception
-{
-    protected $message = 'Área restrita! Você não tem permissão de acesso.';
-}
