--- src/Authentication/JWT/Exceptions/RestrictAccessException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Bludata\Authentication\JWT\Exceptions;
-
-final class RestrictAccessException extends \Exception
-{
-    protected $message = 'Área restrita! Efetue o login antes de continuar.';
-}
