--- src/Authentication/JWT/Interfaces/AuthRepositoryInterface.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Bludata\Authentication\JWT\Interfaces;
-
-interface AuthRepositoryInterface
-{
-}
