+++ src/Lumen/Authentication/JWT/Exceptions/NotPermissionAccessException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Bludata\Lumen\Authentication\JWT\Exceptions;
+
+final class NotPermissionAccessException extends \Exception
+{
+    protected $message = 'Área restrita! Você não tem permissão de acesso.';
+}
