+++ src/Lumen/Authentication/JWT/Interfaces/AuthRepositoryInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Bludata\Lumen\Authentication\JWT\Interfaces;
+
+interface AuthRepositoryInterface
+{
+}
