+++ src/Lumen/Authentication/JWT/Providers/JWTServiceProvider.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Bludata\Lumen\Authentication\JWT\Providers;
+
+use Exception;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\ServiceProvider;
+
+class JWTServiceProvider extends ServiceProvider
+{
+    public function register()
+    {
+        $this->app->bind('Bludata\Lumen\Authentication\JWT\Interfaces\JWTInterface', 'Bludata\Lumen\Authentication\JWT\Libs\JWT');
+
+        $this->app['auth']->viaRequest(
+            'api', function ($request) {
+                if ($token = $request->header('authorization')) {
+                    $auth = app('Bludata\Lumen\Authentication\JWT\Interfaces\AuthRepositoryInterface');
+
+                    try {
+                        $user = $auth->getUserLoggedByToken($token);
+                    } catch (Exception $e) {
+                        abort(401, $e->getMessage());
+                    }
+
+                    return $user;
+                }
+            }
+        );
+    }
+}
