+++ src/Lumen/Providers/CustomConnectionSqlanywhereServiceProvider.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Bludata\LumenProviders;
+
+use Illuminate\Support\ServiceProvider;
+
+class CustomConnectionSqlanywhereServiceProvider extends ServiceProvider
+{
+    public function register()
+    {
+    }
+
+    public function boot()
+    {
+        app('LaravelDoctrine\ORM\Configuration\Connections\ConnectionManager')->extend('sqlanywhere', function () {
+            return config('database.connections.sqlanywhere');
+        });
+    }
+}
