+++ src/Lumen/Services/BaseService.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Bludata\Lumen\Services;
+
+abstract class BaseService
+{
+    /**
+     * @var Bludata\LumenRepositories\BaseRepository
+     */
+    protected $mainRepository;
+
+    public function getMainRepository()
+    {
+        return $this->mainRepository;
+    }
+}
