--- src/Services/BaseService.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Bludata\Services;
-
-abstract class BaseService
-{
-    /**
-     * @var Bludata\Repositories\BaseRepository
-     */
-    protected $mainRepository;
-
-    public function getMainRepository()
-    {
-        return $this->mainRepository;
-    }
-}
